Implemented Course class and its entity relations. fixed a json serialization bug:
I found a json bug that wasn't caught by our tests after adding the person -> RoleSchool one to many relationship. when JSON serializing a personobject it would stack overflow on a cyclical dependency because while serializing a persons role(s), it would find the owner field of the role and try and serialize that, and its roles. and so on and so on.
=========== 
CURRENT FIX:
expose the fields you want GSON so serialize with the @Expose annotation
greate your GSON object with:
Gson gson = new GsonBuilder().excludeFieldsWithoutExposeAnnotation().create();
and serialize with that.